Fix compilation bug

+++ includes/so_long.h
@@ -6,7 +6,7 @@
 /*   By: aborda <aborda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/15 14:29:16 by aborda            #+#    #+#             */
-/*   Updated: 2026/01/29 16:38:08 by aborda           ###   ########.fr       */
+/*   Updated: 2026/01/30 11:10:30 by aborda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ typedef struct s_flood
 	int		nb_collectible;
 	int		found_collectible;
 	int		found_exit;
-	int		player_pos_x;
+	size_t	player_pos_x;
 	int		player_pos_y;
 }	t_flood;
 
@@ -49,7 +49,7 @@ t_flood	*init_s_flood(t_map *map);
 char	**init_map_copy(t_map *map, t_flood *flood);
 int		init_player_pos(t_flood *flood, t_map *map);
 int		init_nb_collectible(t_flood *flood, t_map *map);
-int		flood_fill(t_flood *flood, int player_pos_x, int player_pos_y);
+int		flood_fill(t_flood *flood, size_t player_pos_x, int player_pos_y);
 
 //VALIDATION
 int		is_valid_map(t_map *map);
